requirements.txt
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
Pillow==12.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
reportlab==4.4.10
